+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/api/config/IOperationAbilityConfig.kt
@@ -1,0 +1,7 @@
+package site.siredvin.peripheralium.api.config
+
+interface IOperationAbilityConfig {
+    val isInitialCooldownEnabled: Boolean
+    val initialCooldownSensetiveLevel: Int
+    val cooldownTrasholdLevel: Int
+}