+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/api/peripheral/IPeripheralOperation.kt
@@ -1,6 +1,6 @@
 package site.siredvin.peripheralium.api.peripheral
 
-import site.siredvin.peripheralium.api.IConfigHandler
+import site.siredvin.peripheralium.api.config.IConfigHandler
 
 interface IPeripheralOperation<T> : IConfigHandler {
     val initialCooldown: Int