+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/extra/plugins/ItemStoragePlugin.kt
@@ -3,5 +3,7 @@ package site.siredvin.peripheralium.extra.plugins
 import net.minecraft.world.level.Level
 import site.siredvin.peripheralium.api.storage.Storage
 
-class ItemStoragePlugin(override val storage: Storage, override val level: Level) : AbstractItemStoragePlugin() {
+class ItemStoragePlugin(
+    override val storage: Storage, override val level: Level, override val itemStorageTransferLimit: Int
+) : AbstractItemStoragePlugin() {
 }