+++ projects/forge/src/main/kotlin/site/siredvin/peripheralium/ForgePeripheralium.kt
@@ -5,15 +5,11 @@ import net.minecraft.world.item.Item
 import net.minecraft.world.level.block.Block
 import net.minecraftforge.event.CreativeModeTabEvent
 import net.minecraftforge.eventbus.api.SubscribeEvent
-import net.minecraftforge.fml.ModLoadingContext
 import net.minecraftforge.fml.common.Mod
 import net.minecraftforge.fml.common.Mod.EventBusSubscriber
-import net.minecraftforge.fml.config.ModConfig
-import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent
 import net.minecraftforge.registries.DeferredRegister
 import net.minecraftforge.registries.ForgeRegistries
 import site.siredvin.peripheralium.api.storage.ExtractorProxy
-import site.siredvin.peripheralium.common.configuration.ConfigHolder
 import site.siredvin.peripheralium.forge.ForgeIngredients
 import site.siredvin.peripheralium.forge.ForgePeripheraliumPlatform
 import site.siredvin.peripheralium.storage.ForgeStorageUtils
@@ -32,8 +28,6 @@ object ForgePeripheralium {
 
     init {
         PeripheraliumCore.configure(ForgePeripheraliumPlatform(), ForgeIngredients)
-        val context = ModLoadingContext.get()
-        context.registerConfig(ModConfig.Type.COMMON, ConfigHolder.COMMON_SPEC)
         // Register extract storages
         ExtractorProxy.addStorageExtractor(ForgeStorageUtils::extractStorage)
         val eventBus = MOD_CONTEXT.getKEventBus()